refactor: modernize CMake build system and enforce strict C89 compliance
- Extract static library (flint-vm-lib) to avoid duplicate compilation
- Replace include_directories with target_include_directories
- Use CMAKE_C_STANDARD 90 with CMAKE_C_EXTENSIONS OFF instead of CMAKE_C_FLAGS
- Add -pedantic-errors to core library for strict C89 enforcement
- Remove unused enable_testing()
- Fix C89 violations: mixed declarations, variadic macros, enum trailing comma, void* to function pointer cast
- Refactor logger to use double-paren macro idiom (C89 compatible)
- Use union for void* to function pointer conversion (Lua-style)
- Refactor test framework: add RUN_TEST macro with #func stringification

+++ src/byte_code_printer.c
@@ -28,7 +28,7 @@ void print_byte_code(ByteCodePrinter *printer, Byte *code, size_t code_length) {
       return;
     } else {
       op_name = opcode_info[op][0];
-      fprintf(printer->output, "%zu: ", i);
+      fprintf(printer->output, "%lu: ", (unsigned long)i);
       if (strcmp(opcode_info[op][1], "") == 0) {
         fprintf(printer->output, "%s\n", op_name);
         i++;
